Added required JsonInclude attributes Moved APIAnswer, Tile, GridNodes and Entity in Class/Response folder Renamed PostResponse.cs to AsyncResponse.cs Added PlayerState console line
