--- IAH_SinglePlayerAutomation/Class/GridNode.cs
@@ -1,9 +1,0 @@
-using System.Numerics;
-
-namespace IAH_SinglePlayerAutomation.Class;
-
-public class GridNode
-{
-	public bool blocked;
-	public Vector3 position;
-}
