+++ IAH_SinglePlayerAutomation/Class/Response/TransitionResponse.cs
@@ -1,6 +1,9 @@
+using System.Text.Json.Serialization;
+
 namespace IAH_SinglePlayerAutomation;
 
 public class TransitionResponse
 {
-	public string? state;
+	[JsonInclude]
+	public required string state;
 }
