--- IAH_SinglePlayerAutomation/Class/Tile.cs
@@ -1,16 +1,0 @@
-using System.Numerics;
-
-namespace IAH_SinglePlayerAutomation.Class;
-
-public class Tile
-{
-	public string? equipType;
-
-	public string? frameworkType;
-	public bool isBusy;
-	public bool isOpen;
-	public string? mainType;
-	public Vector3 position;
-	public string? type;
-	public required string uniqueID;
-}
